foundation for cross-crate communication
