+++ shared/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "shared"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+serde = "1.0.210"
+serde_json = "1.0.129"
+tokio = { version = "1", features = ["macros"] }
