+++ standalone/Cargo.toml
@@ -18,6 +18,7 @@ futures = "0.3.31"
 test-log = "0.*"
 base64 = "0.21"
 uuid = "1.17.0"
+shared = { path = "../shared" }
 
 [features]
 neo4j = ["ast/neo4j"]
